requirements.txt
Package                   Version
------------------------- -----------
absl-py                   2.4.0
aiohappyeyeballs          2.6.1
aiohttp                   3.13.3
aiohttp-cors              0.8.1
aiosignal                 1.4.0
altgraph                  0.17.5
astunparse                1.6.3
attrs                     26.1.0
certifi                   2026.2.25
charset-normalizer        3.4.6
colorama                  0.4.6
flatbuffers               25.12.19
frozenlist                1.8.0
gast                      0.7.0
google-pasta              0.2.0
greenlet                  3.3.2
grpcio                    1.78.0
h5py                      3.14.0
hmmlearn                  0.3.3
idna                      3.11
joblib                    1.5.3
keras                     3.13.2
libclang                  18.1.1
markdown-it-py            4.0.0
mdurl                     0.1.2
ml_dtypes                 0.5.4
multidict                 6.7.1
namex                     0.1.0
numpy                     2.4.3
opt_einsum                3.4.0
optree                    0.19.0
packaging                 26.0
pandas                    3.0.1
pefile                    2024.8.26
pip                       26.0.1
propcache                 0.4.1
protobuf                  7.34.1
pyfaidx                   0.9.0.4
Pygments                  2.19.2
pyinstaller               6.19.0
pyinstaller-hooks-contrib 2026.3
python-dateutil           2.9.0.post0
pywin32-ctypes            0.2.3
requests                  2.32.5
rich                      14.3.3
scikit-learn              1.8.0
scipy                     1.17.1
setuptools                82.0.1
six                       1.17.0
SQLAlchemy                2.0.48
tensorflow                2.21.0
termcolor                 3.3.0
threadpoolctl             3.6.0
tqdm                      4.67.3
typing_extensions         4.15.0
tzdata                    2025.3
urllib3                   2.6.3
wheel                     0.46.3
winbbi                    0.1.1
wrapt                     2.1.2
yarl                      1.23.0
